backend-db完成

+++ backend/src/sqls.py
@@ -1,17 +1,31 @@
 import psycopg
 from psycopg.rows import dict_row
+from psycopg import sql
 from typing import List
 from replay import ReplayPost
 import functools
+import secrets
+import hashlib
 
 from contextlib import contextmanager
 import os
 
 from datetime import datetime
 
+STRETCH_NUMBER = 126990
+
 # conn = psycopg.connect()
 # cur = conn.cursor()
 # cur.execute(query)
+
+
+def encrypt_password(raw_password: str, salt: str):
+    return hashlib.pbkdf2_hmac(
+        "sha256",
+        raw_password.encode(),
+        (salt + str(os.getenv("HASH_PEPPER"))).encode(),
+        STRETCH_NUMBER,
+    ).hex()
 
 
 def build_db_kwargs() -> str:
@@ -57,6 +71,7 @@ class SQLReplays:
 
     @staticmethod
     def insert_replay_meta_data(replay_post: ReplayPost):
+        salt = secrets.token_hex()
         with db.transactional() as conn:
             with conn.cursor(row_factory=dict_row) as cur:
                 cur.execute(
@@ -71,7 +86,8 @@ class SQLReplays:
                             game_version,
                             slow_rate,
                             upload_comment,
-                            delete_password
+                            delete_password,
+                            salt
                         )
                         VALUES(
                             %(user_name)s,
@@ -83,7 +99,8 @@ class SQLReplays:
                             %(game_version)s,
                             %(slow_rate)s,
                             %(upload_comment)s,
-                            %(delete_password)s
+                            %(delete_password)s,
+                            %(salt)s
                         )
                         RETURNING replay_id;
                         """,
@@ -97,7 +114,10 @@ class SQLReplays:
                         "game_version": replay_post.replay_meta_data.game_version,
                         "slow_rate": replay_post.replay_meta_data.slow_rate,
                         "upload_comment": replay_post.upload_comment,
-                        "delete_password": replay_post.delete_password,
+                        "delete_password": encrypt_password(
+                            replay_post.delete_password, salt
+                        ),
+                        "salt": salt,
                     },
                 )
 
@@ -107,13 +127,30 @@ class SQLReplays:
                 return returning["replay_id"]
 
     @staticmethod
-    def delete_replay(replay_id):
+    def delete_replay(replay_id, requested_raw_delete_password):
         with db.transactional() as conn:
-            with conn.cursor() as cur:
-                cur.execute("DELETE FROM replays WHERE replay_id = %s;", (replay_id,))
+            with conn.cursor(row_factory=dict_row) as cur:
+                # cur.execute("DELETE FROM replays WHERE replay_id = %s;", (replay_id,))
+                cur.execute(
+                    "SELECT delete_password, salt FROM replays WHERE replay_id=%s",
+                    (replay_id,),
+                )
+
+                selected = cur.fetchone()
+
+                if selected == None:
+                    raise KeyError("DBにリプレイメタデータが存在しませんでした")
+
+                if (
+                    encrypt_password(requested_raw_delete_password, selected["salt"])
+                    != selected["delete_password"]
+                ):
+                    raise ValueError("パスワードが違います")
+
+                cur.execute("DELETE FROM replays WHERE replay_id=%s", (replay_id,))
 
     @staticmethod
-    def select_replay(replay_id) -> ReplayPost:
+    def select_replay(replay_id: int) -> ReplayPost:
         with db.transactional() as conn:
             with conn.cursor(row_factory=dict_row) as cur:
                 cur.execute(
@@ -139,7 +176,6 @@ class SQLReplays:
                     raise ValueError(
                         "リプレイファイルのメタデータを取得できませんでした"
                     )
-                print(selected)
                 replay_post = ReplayPost.new_from_input(
                     replay_id=replay_id,
                     user_name=selected["user_name"],
@@ -154,7 +190,7 @@ class SQLReplays:
                     delete_password=selected["delete_password"],
                 )
                 if cur.fetchone() != None:
-                    raise ValueError("データベースの不整合が起こっています")
+                    raise KeyError("データベースの不整合が起こっています")
                 return replay_post
 
     @staticmethod
@@ -163,37 +199,56 @@ class SQLReplays:
     ) -> List[ReplayPost]:
         with db.transactional() as conn:
             with conn.cursor(row_factory=dict_row) as cur:
-
+                # ASCとDESCを埋め込むことができない為冗長な書き方をしている
                 if is_asc:
-                    order = "ASC"
+                    cur.execute(
+                        sql.SQL(
+                            """
+                            SELECT
+                                replay_id,
+                                user_name,
+                                replay_name,
+                                created_at,
+                                stage,
+                                score,
+                                uploaded_at,
+                                game_version,
+                                slow_rate,
+                                upload_comment,
+                                delete_password
+                            FROM replays
+                            ORDER BY {} ASC LIMIT %s OFFSET %s;
+                        """
+                        ).format(
+                            sql.Identifier(sort),
+                        ),
+                        (limit, offset),
+                    )
                 else:
-                    order = "DESC"
-                print(sort, order)
-                cur.execute(
-                    """
-                        SELECT
-                            replay_id,
-                            user_name,
-                            replay_name,
-                            created_at,
-                            stage,
-                            score,
-                            uploaded_at,
-                            game_version,
-                            slow_rate,
-                            upload_comment,
-                            delete_password
-                        FROM replays
-                        ORDER BY %(sort)s DESC
-                        LIMIT %(limit)s OFFSET %(offset)s;
-                    """,
-                    {
-                        "sort": sort,
-                        # "order": order,
-                        "limit": limit,
-                        "offset": offset,
-                    },
-                )
+                    cur.execute(
+                        sql.SQL(
+                            """
+                            SELECT
+                                replay_id,
+                                user_name,
+                                replay_name,
+                                created_at,
+                                stage,
+                                score,
+                                uploaded_at,
+                                game_version,
+                                slow_rate,
+                                upload_comment,
+                                delete_password
+                            FROM replays
+                            ORDER BY {} DESC LIMIT %s OFFSET %s;
+                        """
+                        ).format(
+                            sql.Identifier(sort),
+                        ),
+                        (limit, offset),
+                    )
+
                 selected = cur.fetchone()
                 if selected == None:
                     raise ValueError(
